re: #275 renamed InotifyBuffer to InotifyMoveEventGrouper
